server/requirements.txt
# FastAPI and server
fastapi>=0.100.0,<0.115.0
uvicorn[standard]>=0.23.0,<0.35.0
python-multipart>=0.0.5,<0.1.0

# Database
sqlmodel>=0.0.14,<0.1.0
sqlalchemy>=2.0.0,<2.1.0

# Authentication
python-jose[cryptography]>=3.3.0,<4.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
bcrypt>=4.0.0,<5.0.0

# Environment
python-dotenv>=1.0.0,<2.0.0

# Validation
pydantic>=2.0.0,<3.0.0
email-validator>=2.0.0,<3.0.0

# Utilities
httpx>=0.25.0,<1.0.0
